Refactoring jobs sub-collection - added under solvers namespace - moved fake implementations away
